+++ src/methods/$collate/async-collate.d.ts
@@ -6,28 +6,13 @@
  * More information can be found in CONTRIBUTING.md
  */
 
-/* tslint:disable unified-signatures */
-import { AsyncSourceIterable, AsyncResultIterable } from "../../types/async-iterable"; // prettier-ignore
-
+import { AsyncSourceIterable, AsyncResultIterable } from '../../types/async-iterable';
 declare function asyncCollate<T>(...sources: Array<AsyncSourceIterable<T>>): AsyncResultIterable<T>;
-declare function asyncCollate<T>(
-  step: number,
-  ...sources: Array<AsyncSourceIterable<T>>
-): AsyncResultIterable<T>;
-declare function asyncCollate<T>(
-  start: number,
-  step: number,
-  ...sources: Array<AsyncSourceIterable<T>>
-): AsyncResultIterable<T>;
 declare function asyncCollate<T>(
   comparator: (a: T, b: T) => number,
   ...sources: Array<AsyncSourceIterable<T>>
 ): AsyncResultIterable<T>;
 declare function asyncCollate<T>(
-  options: {
-    start?: number;
-    step?: number;
-  },
-  ...sources: Array<AsyncSourceIterable<T>>
-): AsyncResultIterable<T>;
+  comparator: (a: T, b: T) => number,
+): (...sources: Array<AsyncSourceIterable<T>>) => AsyncResultIterable<T>;
 export default asyncCollate;